projects/fis-regs-mvp/requirements.txt
streamlit>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pypdf>=3.17.0
python-dotenv>=1.0.0
google-genai>=0.2.0
google-generativeai>=0.3.0
openai>=1.12.0
anthropic>=0.18.0
